Fixed input of time allocations
